# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Examify API",
    "version": "1.1.0",
    "description": "Full API documentation for Examify project, refactored by service"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "User authentication endpoints"
    },
    {
      "name": "Exam",
      "description": "Endpoints for managing exams, questions, and results"
    },
    {
      "name": "Teacher",
      "description": "Endpoints for teachers to manage their classes"
    },
    {
      "name": "Student",
      "description": "Endpoints for students to join and view classes"
    },
    {
      "name": "Class-Management",
      "description": "CRUD operations and general class management"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roleId": {
            "type": "integer"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "Exam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "quantityQuestion": {
            "type": "integer"
          },
          "classId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        }
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "questionId": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "ExamResult": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "score": {
            "type": "number"
          },
          "submitAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "className": {
            "type": "string"
          },
          "classCode": {
            "type": "string"
          },
          "teacherId": {
            "type": "integer"
          }
        }
      },
      "ClassStudentResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "StudentsList": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng ký tài khoản mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "roleId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Đăng nhập",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/create-exam": {
      "post": {
        "tags": [
          "Exam"
        ],
        "summary": "Tạo đề thi mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "integer"
                  },
                  "quantityQuestion": {
                    "type": "integer"
                  },
                  "classId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tạo đề thi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/add-question/{examId}": {
      "post": {
        "tags": [
          "Exam"
        ],
        "summary": "Thêm câu hỏi và đáp án vào đề thi",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "questionList": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "content": {
                          "type": "string"
                        },
                        "score": {
                          "type": "number"
                        },
                        "correctAnswerIndex": {
                          "type": "integer"
                        },
                        "answers": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Answer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm câu hỏi và đáp án thành công"
          }
        }
      }
    },
    "/api/exam/getExamByClass/{classId}": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "Lấy danh sách đề thi theo mã lớp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy đề thi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/getExamDetail/{id}": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "Lấy đề thi theo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy đề thi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/submit-exam/{examId}": {
      "post": {
        "tags": [
          "Exam"
        ],
        "summary": "Học sinh nộp bài thi",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionId": {
                          "type": "integer"
                        },
                        "answerId": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nộp bài thi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam/exam-results/{userId}": {
      "get": {
        "tags": [
          "Exam"
        ],
        "summary": "Lấy kết quả thi của học sinh",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách kết quả thi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/teacher/getClasses/{teacherId}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Lấy tất cả lớp của giáo viên",
        "parameters": [
          {
            "name": "teacherId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách lớp của giáo viên",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/teacher/createClass": {
      "post": {
        "tags": [
          "Teacher"
        ],
        "summary": "Tạo lớp mới (giáo viên)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "teacherId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          }
        }
      }
    },
    "/api/teacher/getStudentsInClass/{classId}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Lấy danh sách học sinh trong lớp",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of students",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentsList"
                }
              }
            }
          }
        }
      }
    },
    "/api/teacher/deleteStudent": {
      "delete": {
        "tags": [
          "Teacher"
        ],
        "summary": "Xóa học sinh khỏi lớp",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassStudentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/join": {
      "post": {
        "tags": [
          "Student"
        ],
        "summary": "Học sinh tham gia lớp bằng mã lớp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "integer"
                  },
                  "classCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student joined class successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassStudentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/student/getClasses/{studentId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Lấy danh sách lớp học sinh tham gia",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of classes student joined",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/class/getAllClass": {
      "get": {
        "tags": [
          "Class-Management"
        ],
        "summary": "Lấy tất cả lớp",
        "responses": {
          "200": {
            "description": "List of classes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/class/{classId}": {
      "get": {
        "tags": [
          "Class-Management"
        ],
        "summary": "Lấy thông tin lớp theo ID",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Class-Management"
        ],
        "summary": "Cập nhật lớp",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "teacherId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Class updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}